ui objects refactoring
