+++ JDI.Light/JDI.Light.Tests/UIObjects/Pages/ContactPage.cs
@@ -2,6 +2,7 @@
 using JDI.Light.Elements.Common;
 using JDI.Light.Elements.Composite;
 using JDI.Light.Interfaces.Common;
+using JDI.Light.Tests.UIObjects.Forms;
 using JDI.Light.Tests.UIObjects.Sections;
 
 namespace JDI.Light.Tests.UIObjects.Pages
